Improved cave::HashMap Iter time using linked list
Previously cave::HashMap iteration time was about 30 times slower than std::unordered_map. So I changed it in a way that each element also stores a pointer to its previous and next element (order is based on addition order, but you should not rely on the order!), making it easy to iterate (as it was a linked list). It's now about 2 times slower than std::unordered_map. Still bad, but at least it's now "usable". :+1: